@@ -23,6 +23,7 @@ module.exports = {
   ],
   rules: {
     "react/prop-types": "off",
-    "react/jsx-one-expression-per-line": "off"
+    "react/jsx-one-expression-per-line": "off",
+    "linebreak-style": "off",
   },
 };
